the pairwise record's version byte compares as an int, because kotlin refuses a byte against a literal and the ci runner was the only compiler that saw it

+++ clients/android/core/src/main/kotlin/com/migo/core/store/SessionStore.kt
@@ -359,7 +359,7 @@ class SessionStore private constructor(
      */
     private fun decodePairwise(stored: ByteArray): StoredSession {
         if (stored.isEmpty()) throw SessionStoreError.Unreadable
-        if (stored[0] != PAIRWISE_ENTRY_VERSION) {
+        if (stored[0].toInt() != PAIRWISE_ENTRY_VERSION) {
             // A record from before the secret was retained: the whole body is the snapshot, and
             // `restore` enforces the snapshot's own version byte.
             return StoredSession(RatchetSession.restore(stored), null)
